refactor: reorg ui components
