+++ web/src/components/structure/Navbar.tsx
@@ -1,10 +1,10 @@
 import type { Session } from "next-auth";
 import { signIn } from "next-auth/react";
 import Link from "next/link";
+import Menu from "~/components/Menu";
+import WidgetSidebar from "~/components/WidgetSidebar";
+import { Button } from "~/components/ui/button";
 import { APP_NAME } from "~/lib/basic/const";
-import Menu from "./Menu";
-import WidgetSidebar from "./WidgetSidebar";
-import { Button } from "./ui/button";
 
 type Props = {
   sesh: Session | null;
