+++ web/src/pages/w/[id].tsx
@@ -2,8 +2,8 @@ import type { NextPage } from "next";
 import { useRouter } from "next/router";
 import { useState, type ChangeEvent } from "react";
 import { getNameForWidgetType } from "~/application/widget.service";
-import Separator from "~/components/Separator";
-import SiteWrapper from "~/components/SiteWrapper";
+import Separator from "~/components/structure/Separator";
+import SiteWrapper from "~/components/structure/SiteWrapper";
 import Textarea from "~/components/Textarea";
 import { Button } from "~/components/ui/button";
 import { api } from "~/lib/api/api";
